# make has_edge a bit faster 