--- a/src/utils.jl
+++ b/src/utils.jl
@@ -57,7 +57,7 @@ function insorted(item, collection; rev=false)
     return !isempty(index)
 end
 
-function insorted(x::T, A::AbstractVector{T}) where T
+@inline function insorted(x::T, A::AbstractVector{T}) where T
     n = length(A)
     lo = 0
     hi = n+1