About 10-15% speedup 

**Old code**
```julia
julia> using LightGraphs, SNAPDatasets, BenchmarkTools

julia> g = loadsnap(:amazon0302)
{262111, 1234877} directed simple UInt32 graph

julia> function f(g::AbstractGraph{T}, n::Int64 = 10000) where T
           nvg = nv(g)
           srcs = rand(T(1):T(nvg), n)
           dsts = rand(T(1):T(nvg), n)
           for i = 1:n
               has_edge(g, srcs[i], dsts[i])
           end
       end
f (generic function with 2 methods)

julia> @btime f(g, 200000)
  30.691 ms (4 allocations: 1.53 MiB)

julia> g = loadsnap(:soc_slashdot0902_u)
{82168, 582533} undirected simple Int64 graph

julia> @btime f(g, 200000)
  27.838 ms (4 allocations: 3.05 MiB)
```
**New code**
```julia
julia> using LightGraphs, SNAPDatasets, BenchmarkTools

julia> g = loadsnap(:amazon0302)
{262111, 1234877} directed simple UInt32 graph

julia> function f(g::AbstractGraph{T}, n::Int64 = 10000) where T
           nvg = nv(g)
           srcs = rand(T(1):T(nvg), n)
           dsts = rand(T(1):T(nvg), n)
           for i = 1:n
               has_edge(g, srcs[i], dsts[i])
           end
       end
f (generic function with 2 methods)

julia> @btime f(g, 200000)
  25.885 ms (4 allocations: 1.53 MiB)

julia> g = loadsnap(:soc_slashdot0902_u)
{82168, 582533} undirected simple Int64 graph

julia> @btime f(g, 200000)
  22.702 ms (4 allocations: 3.05 MiB)
```