@sbromberger 2020-07-15
So my question is why our `insorted` is faster than Julia's built-in one, and whether this deserves a PR to Julia.