@abhinavmehndiratta 2020-07-15
> So my question is why our `insorted` is faster than Julia's built-in one, and whether this deserves a PR to Julia.

Our version is just inlined and the rest is exactly the same, https://github.com/JuliaLang/julia/blob/44fa15b1502a45eac76c9017af94332d4557b251/base/sort.jl#L177 you can see that we use the exact same code here.