@simonschoelly 2020-07-15
Weird that we have some tests failing.

I can imagine that a specialised version might indeed be slightly faster than the builtin one - we know that we deal here with only positive integers.