@abhinavmehndiratta 2020-07-15
> Weird that we have some tests failing.
> 
> I can imagine that a specialised version might indeed be slightly faster than the builtin one - we know that we deal here with only positive integers.

The build failure is unrelated. I dont think we'll have any significant advantage with the fact that we only have positive integers, since binary search only relies on comparison.