@abhinavmehndiratta 2020-07-16
I've fixed `insorted` for reverse order and this is ready to merge. Benchmarks are almost unaffected -
```julia
julia> g = loadsnap(:amazon0302)
{262111, 1234877} directed simple UInt32 graph

julia> function f(g::AbstractGraph{T}, n::Int64 = 10000) where T
           nvg = nv(g)
           srcs = rand(T(1):T(nvg), n)
           dsts = rand(T(1):T(nvg), n)
           for i = 1:n
               has_edge(g, srcs[i], dsts[i])
           end
       end
f (generic function with 2 methods)

julia> @btime f(g, 200000)
  24.734 ms (4 allocations: 1.53 MiB)

julia> g = loadsnap(:soc_slashdot0902_u)
{82168, 582533} undirected simple Int64 graph

julia> @btime f(g, 200000)
  22.064 ms (4 allocations: 3.05 MiB)
```