@sbromberger 2020-07-16
why are we getting inference errors in nightly? I thought we fixed that.