@abhinavmehndiratta 2020-07-16
> why are we getting inference errors in nightly? I thought we fixed that.

No we didn't. I think you opened an issue for it but it isn't fixed